% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball_visuals.R, R/boxscores.R
\name{check_pitcherid}
\alias{check_pitcherid}
\title{Helper function that checks whether the pitcherid provided is valid in the given dataset}
\usage{
check_pitcherid(data, pitcherid)

check_pitcherid(data, pitcherid)
}
\arguments{
\item{data}{trackman baseball dataset}

\item{pitcherid}{The ID of pitcher}
}
\value{
Stop if pitcherid is not valid

Stop if pitcherid is not valid
}
\description{
Helper function that checks whether the pitcherid provided is valid in the given dataset

Helper function that checks whether the pitcherid provided is valid in the given dataset
}
