% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitcher_summary.R
\name{get_pitching_summary}
\alias{get_pitching_summary}
\title{Helper function to summarize columns of interest for a pitcher}
\usage{
get_pitching_summary(pitcher_data)
}
\arguments{
\item{pitcher_data}{A trackman baseball dataframe, subsetted by pitcher of
interest with PitchCallClass column to calculate strikes}
}
\value{
A dataframe with pitch type as columns and summarized values as rows
}
\description{
Helper function to summarize columns of interest for a pitcher
}
