% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitcher_performance.R
\name{helper_function}
\alias{helper_function}
\title{Helper function description}
\usage{
helper_function(pitcher_data)
}
\arguments{
\item{name}{description}
}
\description{
Helper function description
}
