% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball_visuals.R
\name{location_chart}
\alias{location_chart}
\title{Helper function that creates a pitch location plot}
\usage{
location_chart(data, firstname, lastname, game_date)
}
\arguments{
\item{data}{trackman baseball dataset}

\item{pitcherid}{The ID of pitcher}
}
\value{
A scatter plot displaying pitch locations by pitch type
}
\description{
Helper function that creates a pitch location plot
}
