% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball_visuals.R
\name{movement_chart}
\alias{movement_chart}
\title{Helper function that creates a movement plot}
\usage{
movement_chart(data, firstname, lastname, game_date)
}
\arguments{
\item{data}{trackman baseball dataset}

\item{pitcherid}{The ID of pitcher}
}
\value{
A scatter plot displaying pitch movement profiles by pitch type
}
\description{
Helper function that creates a movement plot
}
