% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxscores.R
\name{pitcher_box_score}
\alias{pitcher_box_score}
\title{Function to provide a pitcher's box score from Trackman data}
\usage{
pitcher_box_score(data, pitcherid)
}
\arguments{
\item{data}{A Trackman baseball dataset}

\item{pitcherid}{The numeric ID of the pitcher to summarize}
}
\value{
A box score data frame
}
\description{
Function to provide a pitcher's box score from Trackman data
}
