% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball_visuals.R
\name{pitcher_chart}
\alias{pitcher_chart}
\title{Function to create plots from a pitcher's Trackman data}
\usage{
pitcher_chart(data, pitcherid, type = "movement")
}
\arguments{
\item{data}{trackman baseball dataset}

\item{pitcherid}{The ID of pitcher}

\item{type}{type of chart (movement, release, location)}
}
\value{
A scatter plot displaying desired pitch data specified by the chart type
}
\description{
Function to create plots from a pitcher's Trackman data
}
