% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitcher_performance.R
\name{pitcher_performance}
\alias{pitcher_performance}
\title{Function to summarize a pitcher's performance in a baseball game}
\usage{
pitcher_performance(data, pitcherid)
}
\arguments{
\item{data}{A trackman baseball dataset}

\item{pitcherid}{The numeric ID of the pitcher to summarize}
}
\value{
A dataframe summarizing pitcher performance for a specified pitcher
in one baseball game, counting up innings, strikeouts, walks, and runs.
}
\description{
Function to summarize a pitcher's performance in a baseball game
}
