% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitcher_summary.R
\name{pitcher_summary}
\alias{pitcher_summary}
\title{Function to summarize a pitcher's baseball game statistics}
\usage{
pitcher_summary(data, pitcherid, type = "tagged")
}
\arguments{
\item{data}{A trackman baseball dataset}

\item{pitcherid}{The numeric ID of the pitcher to summarize}

\item{pitch_type}{The type of pitch classification to use; "tagged" or "auto".
Default value is "tagged".}
}
\value{
A table indicating pitching statistics for a specified pitcher
in one baseball game, both overall and grouped by pitch type.
}
\description{
Function to summarize a pitcher's baseball game statistics
}
