% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball_visuals.R
\name{release_chart}
\alias{release_chart}
\title{Helper function that creates a release point plot}
\usage{
release_chart(data, firstname, lastname, game_date)
}
\arguments{
\item{data}{trackman baseball dataset}

\item{pitcherid}{The ID of pitcher}
}
\value{
A scatter plot displaying pitcher's release points by pitch type
}
\description{
Helper function that creates a release point plot
}
