% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball_visuals.R, R/boxscores.R
\name{split_pitcher_name}
\alias{split_pitcher_name}
\title{Helper function that splits the pitcher's name into "Last" "First"}
\usage{
split_pitcher_name(data)

split_pitcher_name(data)
}
\arguments{
\item{data}{pitcher dataset}
}
\value{
Two strings: "Last" "First"

Two strings: "Last" "First"
}
\description{
Helper function that splits the pitcher's name into "Last" "First"

Helper function that splits the pitcher's name into "Last" "First"
}
